+++ GCIEvaluator/src/main/java/de/tudresden/inf/tcs/OntException.java
@@ -1,0 +1,11 @@
+
+public class OntException extends Exception {
+
+	public OntException(String arg0) {
+		super(arg0);
+	}
+
+	public OntException(Throwable arg0) {
+		super(arg0);
+	}
+}
