--- Ontol/.gitignore
@@ -1,1 +1,0 @@
-/test.class
