--- Ontol/eval/.gitignore
@@ -1,3 +1,0 @@
-/EvaluatingGCIs.class
-/OWLVetoesListener.class
-/TestingOWLAPI.class
