--- Ontol/eval/OntException.java
@@ -1,11 +1,0 @@
-
-public class OntException extends Exception {
-
-	public OntException(String arg0) {
-		super(arg0);
-	}
-
-	public OntException(Throwable arg0) {
-		super(arg0);
-	}
-}
